API : fix roletype repo
